python-playwright-enterprise-framework/requirements.txt
pytest
playwright
pytest-playwright
pytest-xdist
allure-pytest
python-dotenv
PyYAML
